Decompose the unified chat monolith into focused hooks; serve it at /
The 2,635-line unified page becomes a ~620-line composition over seven
hooks (chain controller, thread persistence, codex tasks, branches,
assistant tasks, doc read, finder) plus shared helpers. Zero behavior
change, verified by build + manual flow review.

The unified experience is the product, so it now lives at the root route;
/demos/unified redirects (preserving ?chatId) and the standalone feature
demos remain under /demos/*.
